+++ src/Messenger/EventStoreTransactionMiddleware.php
@@ -42,16 +42,11 @@ final class EventStoreTransactionMiddleware implements MiddlewareInterface
             $this->eventStore->rollback();
 
             if ($e instanceof HandlerFailedException) {
-                $method = 'getWrappedExceptions';
-                if (! \method_exists($e, $method)) {
-                    $method = 'getNestedExceptions';
-                }
-
                 // Remove all HandledStamp from the envelope so the retry will execute all handlers again.
                 // When a handler fails, the queries of allegedly successful previous handlers just got rolled back.
                 throw new HandlerFailedException(
                     $e->getEnvelope()->withoutAll(HandledStamp::class),
-                    $e->$method()
+                    $e->getWrappedExceptions()
                 );
             }