+++ test/DependencyInjection/Compiler/ProjectionOptionsPassTest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace ProophTest\Bundle\EventStore\DependencyInjection\Compiler;
 
+use PHPUnit\Framework\Attributes\Test;
 use Prooph\Bundle\EventStore\DependencyInjection\Compiler\ProjectionOptionsPass;
 use Prooph\Bundle\EventStore\DependencyInjection\ProophEventStoreExtension;
 use Prooph\Bundle\EventStore\Exception\RuntimeException;
@@ -37,7 +38,7 @@ final class ProjectionOptionsPassTest extends CompilerPassTestCase
         $this->registerEmptyServiceLocator('prooph_event_store.projection_options_locator');
     }
 
-    /** @test */
+    #[Test]
     public function it_does_not_allow_inappropriate_classes(): void
     {
         $this->registerProjectionOptions(
@@ -55,7 +56,7 @@ final class ProjectionOptionsPassTest extends CompilerPassTestCase
         $this->compile();
     }
 
-    /** @test */
+    #[Test]
     public function it_does_not_allow_tags_without_projection_name(): void
     {
         $this->registerProjectionOptions(
@@ -72,7 +73,7 @@ final class ProjectionOptionsPassTest extends CompilerPassTestCase
         $this->compile();
     }
 
-    /** @test */
+    #[Test]
     public function it_registers_tagged_projection_options(): void
     {
         $this->registerProjectionOptions(