+++ test/DependencyInjection/Fixture/config/php/unconfigured.php
@@ -1,3 +1,5 @@
+<?php
+
 /**
  * This file is part of prooph/event-store-symfony-bundle.
  * (c) 2014-2024 Alexander Miertsch <kontakt@codeliner.ws>
@@ -6,3 +8,11 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
+declare(strict_types=1);
+
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+
+return static function (ContainerConfigurator $container): void {
+    $container->extension('prooph_event_store', []);
+};